adjust in `Replicates` field

+++ priv/repo/migrations/20220109201031_add_plant_traits.exs
@@ -22,7 +22,7 @@ defmodule Chirinola.Repo.Migrations.AddPlantTraits do
       add :ValueKindName, :string
       add :OrigUncertaintyStr, :string
       add :UncertaintyName, :string
-      add :Replicates, :integer
+      add :Replicates, :float
       add :StdValue, :float
       add :UnitName, :string
       add :RelUncertaintyPercent, :float
